Use generated unique item IDs
Not using data hash as unique key allows having duplicate items which
can cause problems when there are spurious clipboard change events but
can be useful in the future (for example: clone item and edit it).

The unique ID is used for ordering items instead of creation date/time
which is better when quickly adding multiple items (for example using
"add" command).